+++ examples/js/shopping-list-redux-hoc/config.dev.js
@@ -6,6 +6,7 @@ const devConfig = {
     options: {
       host: declareRuntimeEnv('HOST', 'localhost'),
       port: declareRuntimeEnv('PORT', '3000'),
+      emulateWorkerLifetimeLimit: 240000,
     },
   },
   eventstoreAdapter: {