+++ packages/core/core/src/types/runtime.ts
@@ -119,6 +119,17 @@ export type ReplicationState = {
   successEvent: OldEvent | null
 }
 
+export type EventStoreDescription = {
+  eventCount: number
+  secretCount: number
+  setSecretCount: number
+  deletedSecretCount: number
+  isFrozen: boolean
+  lastEventTimestamp: number
+  cursor?: Cursor
+  resourceNames?: { [key: string]: string }
+}
+
 export type Eventstore = {
   saveEvent: (event: Event) => Promise<StoredEventPointer>
   saveSnapshot: (snapshotKey: string, content: string) => Promise<void>
@@ -167,6 +178,8 @@ export type Eventstore = {
   setReplicationPaused: (pause: boolean) => Promise<void>
   getReplicationState: () => Promise<ReplicationState>
   resetReplication: () => Promise<void>
+
+  describe: () => Promise<EventStoreDescription>
 }
 
 export type AggregateMeta = {