+++ packages/runtime/adapters/eventstore-adapters/eventstore-base/src/create-adapter.ts
@@ -98,7 +98,12 @@ const createAdapter = <
     getEventLoaderNative,
   }: AdapterFunctions<ConnectedProps, ConnectionDependencies, Config>,
   connectionDependencies: ConnectionDependencies,
-  options: Config
+  options: Config,
+  prepare?: (
+    props: AdapterPoolPossiblyUnconnected<ConnectedProps>,
+    config: Config,
+    dependencies: ConnectionDependencies
+  ) => void
 ): Adapter => {
   const log: LeveledDebugger & debug.Debugger = getLog(`createAdapter`)
   const config: Config = { ...options }
@@ -143,6 +148,9 @@ const createAdapter = <
     ...primalProps,
     ...emptyProps,
   }
+  if (prepare !== undefined) {
+    prepare(adapterPool, config, connectionDependencies)
+  }
 
   const connectedProps: AdapterPoolPrivateConnectedProps = {
     injectEvent: wrapMethod(adapterPool, injectEvent),
@@ -162,7 +170,12 @@ const createAdapter = <
     waitConnect: wrapMethod(adapterPool, emptyFunction),
     shapeEvent,
     getEventLoaderNative: getEventLoaderNative
-      ? wrapMethod(adapterPool, getEventLoaderNative)
+      ? async (filter) => {
+          return getEventLoaderNative(
+            adapterPool as AdapterPoolConnected<ConnectedProps>,
+            filter
+          )
+        }
       : getEventLoaderNative,
   }
 
@@ -225,7 +238,12 @@ const createAdapter = <
             return
           }
         : establishTimeLimit.bind(null, adapterPool),
-    getEventLoader: wrapMethod(adapterPool, getEventLoader),
+    getEventLoader: async (filter) => {
+      return getEventLoader(
+        adapterPool as AdapterPoolConnected<ConnectedProps>,
+        filter
+      )
+    },
   }
 
   Object.assign<AdapterPoolPossiblyUnconnected<ConnectedProps>, Adapter>(