+++ packages/runtime/adapters/eventstore-adapters/eventstore-postgresql-serverless/src/__mocks__/aws-sdk/clients/rdsdataservice.ts
@@ -1,8 +1,0 @@
-const RDSDataService = jest.fn()
-
-const promised = (): Function =>
-  jest.fn().mockReturnValue({ promise: () => Promise.resolve() })
-
-RDSDataService.prototype.executeStatement = promised()
-
-export default RDSDataService