+++ tests/custom-readmodel-sample/projection.js
@@ -4,10 +4,10 @@ const projection = {
     await store.set(0)
   },
   INCREMENT: async (store, event) => {
-    await store.set((await store.get()) + event.payload)
+    await store.set((await store.get()) + event.payload.count)
   },
   DECREMENT: async (store, event) => {
-    await store.set((await store.get()) - event.payload)
+    await store.set((await store.get()) - event.payload.count)
   },
 }